backend/requirements.txt
Django>=3.0.8,<3.1.0
djangorestframework>=3.11.0,<3.12.0

PyGithub>=1.51,<1.60
psycopg2-binary>=2.8.5,<2.9.0

flake8>=3.8.3,<3.9.0